Create Возведение в степень.Math.pow().7.Count cubes in a Menger Sponge
An illustration of the iterative construction of a Menger sponge
A method of constructing a Menger Sponge can be visualized as follows:

Start from a cube (first part of image).
Scale down the cube so that side length is 1/3 of its original, and make 20 copies of it.
Place the copies so that they measure the same size as the original cube but without its central parts (next part of image)
Repeat the process from step 2 for the new smaller cubes from the previous step.
In each iteration (e.g. repeating the last three steps), the effect will be that parts of the cube will be removed, they'll never be added. Menger sponge will always consist of parts will never be removed, regardless of how many iterations you do.
An alternative explanation:

Start from a cube (first part of image).
Devide each cube into 27 equal sized cubes.
Remove the middle-cube and the six cubes on each side of the group of 27 cubes (second part of image).
Repeat the process from step 2 for the smaller cubes (third and fourth part of image).

Task
In this kata you will create a function that takes non negative integers (from 0 to n) and return the amount of cubes that the Menger Sponge would have in that specific iteration.



Иллюстрация итеративной конструкции губки Менгера
Способ построения губки Менгера можно представить следующим образом:

Начните с куба (первая часть изображения).
Уменьшите размер куба так, чтобы длина его стороны составляла 1/3 от оригинала, и сделайте 20 копий.
Поместите копии так, чтобы они имели тот же размер, что и исходный куб, но без его центральных частей (следующая часть изображения)
Повторите процесс из шага 2 для новых кубиков меньшего размера из предыдущего шага.
В каждой итерации (например, повторяя последние три шага), результатом будет то, что части куба будут удалены, они никогда не будут добавлены. Губка Менгера всегда будет состоять из частей, которые никогда не будут удалены, независимо от того, сколько итераций вы сделаете.
Альтернативное объяснение:

Начните с куба (первая часть изображения).
Разделите каждый куб на 27 кубиков одинакового размера.
Удалите средний куб и шесть кубиков с каждой стороны группы из 27 кубиков (вторая часть изображения).
Повторите процесс из шага 2 для меньших кубов (третья и четвертая части изображения).